ODY-48 updated droplet and authorizedUser types, and made favorited droplet tiles
