--- backend/src/api/favorite/content-types/favorite/schema.json
@@ -1,28 +1,0 @@
-{
-  "kind": "collectionType",
-  "collectionName": "favorites",
-  "info": {
-    "singularName": "favorite",
-    "pluralName": "favorites",
-    "displayName": "Favorite",
-    "description": ""
-  },
-  "options": {
-    "draftAndPublish": true
-  },
-  "pluginOptions": {},
-  "attributes": {
-    "user": {
-      "type": "relation",
-      "relation": "manyToOne",
-      "target": "api::authorized-user.authorized-user",
-      "inversedBy": "favorites"
-    },
-    "droplet": {
-      "type": "relation",
-      "relation": "manyToOne",
-      "target": "api::droplet.droplet",
-      "inversedBy": "favorites"
-    }
-  }
-}
