--- backend/src/api/favorite/controllers/favorite.ts
@@ -1,7 +1,0 @@
-/**
- * favorite controller
- */
-
-import { factories } from '@strapi/strapi'
-
-export default factories.createCoreController('api::favorite.favorite' as any);
