--- backend/src/api/favorite/routes/favorite.ts
@@ -1,7 +1,0 @@
-/**
- * favorite router
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreRouter('api::favorite.favorite' as any);
