--- backend/src/api/favorite/services/favorite.ts
@@ -1,7 +1,0 @@
-/**
- * favorite service
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreService('api::favorite.favorite' as any);
